+++ tests/content.test.ts
@@ -843,6 +843,35 @@ describe('rikaiContent:text search', () => {
     assertTextResultEqual(result, 'せん', seNode, 0, seNode, 1, nNode, 1);
   });
 
+  it('should traverse okurigana in inline-block elements too', () => {
+    // YouTube annotates okurigana inline-block spans.
+    //
+    // See https://github.com/birtles/rikaichamp/issues/535
+    testDiv.innerHTML =
+      '<p><ruby><span>疲</span><rt>つか</rt></ruby><span style="display: inline-block">れた</span></p>';
+
+    const kanjiNode = testDiv.firstChild!.firstChild!.firstChild!
+      .firstChild as Text;
+    const okuriganaNode = testDiv.firstChild!.childNodes[1].firstChild as Text;
+    const bbox = getBboxForOffset(kanjiNode, 0);
+
+    const result = subject.getTextAtPoint({
+      x: bbox.left,
+      y: bbox.top + bbox.height / 2,
+    });
+
+    assertTextResultEqual(
+      result,
+      '疲れた',
+      kanjiNode,
+      0,
+      kanjiNode,
+      1,
+      okuriganaNode,
+      2
+    );
+  });
+
   it('should find text in SVG content', () => {
     testDiv.innerHTML = '<svg><text y="1em">あいうえお</text></svg>';
     const textNode = testDiv.firstChild!.firstChild!.firstChild as Text;